Added a resources directory with migrations to deploy contracts for testing and development purposes; added scripts to package those contracts and drop them in the node_modules directory as symlinks to be used in the dapp source code without modifications.
Modified import statements in the source code to work with separate abi
and networks json files for blockchain contracts and token contracts.

Changed references for Registry contracts to AlphaRegistry.

Fixed all links to etherscan so that the href is undefined if the
current network name does not match a network that is in our file:
before if you tried to run the dapp with any unknown network
(ganache-cli) it would simply crash.

Added a dependency that was missing from package.json (underscore).

Updated dependencies that had vulnerabilities in them.